algorithms/src/test/java/ps_1301_1400/alg_1315_sum_of_nodes_with_even_valued_grandparent/SolutionTest.java
package ps_1301_1400.alg_1315_sum_of_nodes_with_even_valued_grandparent;

import common.TreeNode;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class SolutionTest {
    private final Solution solution = new Solution();

    @Test
    public void testcase1() {
        TreeNode root = new TreeNode(6);
        root.left = new TreeNode(7);
        root.right = new TreeNode(8);

        root.left.left = new TreeNode(2);
        root.left.right = new TreeNode(7);
        root.right.left = new TreeNode(1);
        root.right.right = new TreeNode(3);

        root.left.left.left = new TreeNode(9);
        root.left.right.left = new TreeNode(1);
        root.left.right.right = new TreeNode(4);
        root.right.right.right = new TreeNode(5);

        assertEquals(18, solution.sumEvenGrandparent(root));
    }

}